Update .env File Flow (#5)
This Pull Request introduces `.env` files to be used for production workflows.

The project's `.gitignore` has `.env*.local` as an option, but there were no instructions in the README for using such files.

I believe this project to be using NextAuth for its intended "KeyCloak" auth (which will in fact be using an [OpenIDConnect](https://developer.okta.com/blog/2019/01/23/nobody-cares-about-oauth-or-openid-connect) flow). I have introduced [example values](https://next-auth.js.org/providers/keycloak#example) to be replaced with the correct ones, supplied via a not-public forum, as the values are secret.

Instructions have been added to the README file to make sure the process is obvious and simple for anyone running this project.

Reviewed-on: https://codeberg.org/reality-hack-inc/realityhack.world_frontEnd/pulls/5
Co-authored-by: shane <shane@mitrealityhack.com>
Co-committed-by: shane <shane@mitrealityhack.com>