add admin teams page (#45)
Add teams page, and add ability to edit individual teams.

To test, go to admin, then go to the new Teams card.

Should see teams load, you should see team name, # of members and other top level info. On clicking save, you should see a snackbar confirmation and on reload see the changes are persistent. Reset will revert to the original before last save.

At an individual team page, you will be able to change the team name, team table, and attendees. Note that I do not get any other info from the table, so I only show the id. Will likely need to add a serializer to the backend so we can get location info included in the `/tables/` api endpoint, so we can use the location info in the dropdown autocomplete. Will be hard to add a table if you do not know where it is located. Will need to discuss with @singlerider if this would have detrimental impact on performance.

![image](/attachments/379a6dfc-b69c-4cfe-8ef5-da1f76d6e244)

Reviewed-on: https://codeberg.org/reality-hack-inc/realityhack.world-frontend/pulls/45
Reviewed-by: sdimaren <sdimaren@noreply.codeberg.org>
Co-authored-by: Peter Iordanov <peter@bulbyte.com>
Co-committed-by: Peter Iordanov <peter@bulbyte.com>